backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
motor>=3.7.0
mongomock-motor>=0.0.36
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
passlib>=1.7.4
bcrypt>=4.1.0
typing_extensions
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
apscheduler>=3.10.0
pytz>=2023.3
tzlocal>=5.0
